examples: add a am7xxx-play example program
am7xxx-play is a simple player based on ffmpeg/libav which displays the
output on a am7xxx device. It can be used to capture the desktop screen,
or video from a v4l2 device, or from any other device supported by
libavdevice including video files.

However it is to note that am7xxx-play displays the output as fast as it
can so it is not reliable when used to display video files directly.