am7xxx-play: check if strtok_r is available
On Windows systems, where strtok_r is not available, strtok_s could be
used but this needs the C Run-Time library (msvcrt.dll or msvcr80.dll)
and we can't distribute it.

So for now, when strtok_r is not available, just print a message to warn
the user that the '-o' options is not available.